beta

+++ astarte-core/src/main/java/com/github/harbby/astarte/ResultTask.java
@@ -33,7 +33,7 @@ public class ResultTask<E, R>
     }
 
     @Override
-    public long getTaskId()
+    public int getTaskId()
     {
         return partition.getId();
     }
