+++ astarte-core/src/main/java/com/github/harbby/astarte/ShuffleMapTask.java
@@ -26,7 +26,7 @@ public class ShuffleMapTask<E>
     }
 
     @Override
-    public long getTaskId()
+    public int getTaskId()
     {
         return partition.getId();
     }
